honor visible_only for recent view
